move rapierMath to its own class
